cloud-2022/create_order_function/requirements.txt
azure-functions
psycopg2-binary
azure-servicebus
azure-storage-blob
pandas


